moved api logic into a seperate service
