+++ app/controllers/users_controller.rb
@@ -8,7 +8,7 @@ class UsersController < ApplicationController
   def get_gists
     @user = User.find_by_githubname(params[:slug])
     count = @user.user_gists.count
-    @user.get_gists
+    Github::Gists::Api.new(@user).call_api
     new_gists_count = (@user.user_gists.count - count).positive? ? (@user.user_gists.count - count).positive? : 0
 
     redirect_to user_user_gists_path(user_slug: @user.githubname, new_gists_count: new_gists_count)
